Revert unnecessary change
