+++ FFmpegInterop/FFmpegInterop.vcxproj
@@ -171,6 +171,7 @@
     <ClCompile Include="pch.cpp">
       <PrecompiledHeader>Create</PrecompiledHeader>
     </ClCompile>
+    <ClCompile Include="Utility.cpp" />
     <ClCompile Include="VFWSampleProvider.cpp" />
   </ItemGroup>
   <ItemGroup>
